==> Notebooks/2_EDA_Informacion_Financiera.ipynb <==
# 2 - EDA DATASET INFORMACION FINANCIERA DE LA SUPERINTENDENCIA

**AUTOR: Fabrizio Ramirez Cutimbo**

**OBJETIVO:**

+ Transformar y limpiar los dataset de "Información Financiera" que expone la Superintendencia de Compañías, Valores y Seguros del año 2022 y 2023.
---

import pandas as pd
import numpy as np

## PROCESAMIENTO Información Financiera 2023

### Cargar Dataset

file_path = "Datasets\\estadosFinancieros_2023\\balances_2023_1.txt"
# Lee el archivo usando el delimitador de tabulaciones
pd_dataset = pd.read_csv(file_path, delimiter='\t', dtype={'RUC': 'object', 'EXPEDIENTE': 'object'}, encoding='latin1')

### Forma, Columnas y Datos Generales

pd_dataset.sample(1)

print(f"Dimensiones del conjunto de datos: \n Filas: {pd_dataset.shape[0]}  \n Columnas: {pd_dataset.shape[1]}")

pd_dataset.columns

pd_dataset.isna().sum()

pd_dataset.dtypes

### Selección columnas de interés

pd_dataset_filtered = pd_dataset[['EXPEDIENTE', 'RUC',
                                  'CUENTA_1', 'CUENTA_101', 'CUENTA_10103', 'CUENTA_102',
                                  'CUENTA_2', 'CUENTA_201', 'CUENTA_202', 
                                  'CUENTA_3', 
                                  'CUENTA_401', 'CUENTA_402', 'CUENTA_403', 
                                  'CUENTA_501', 'CUENTA_502', 
                                  'CUENTA_600', 'CUENTA_602', 'CUENTA_603', 'CUENTA_607' ]]

pd_dataset_filtered.sample(1)

# Calcular el porcentaje de valores nulos por columna
pd_dataset_filtered.isnull().mean() * 100

### Transformación a datos numéricos

num_columns = ['CUENTA_1', 'CUENTA_101', 'CUENTA_10103', 'CUENTA_102', 
                'CUENTA_2', 'CUENTA_201', 'CUENTA_202', 
                'CUENTA_3', 
                'CUENTA_401', 'CUENTA_402', 'CUENTA_403', 
                'CUENTA_501','CUENTA_502', 
                'CUENTA_600', 'CUENTA_602', 'CUENTA_603', 'CUENTA_607' ]

pd_dataset_filtered[num_columns] = pd_dataset_filtered[num_columns].apply(
    lambda col: pd.to_numeric(col.str.replace('.', '', regex=False).str.replace(',', '.', regex=False))
)

pd_dataset_filtered.sample(1)

### Renombrar Columnas

diccionario_2023 = {
    'CUENTA_1': 'ACTIVO_2023',
    'CUENTA_101': 'ACTIVO_CORRIENTE_2023',
    'CUENTA_10103': 'INVENTARIOS_2023',
    'CUENTA_102': 'ACTIVOS_NO_CORRIENTES_2023',
    'CUENTA_2': 'PASIVO_2023',
    'CUENTA_201': 'PASIVO_CORRIENTE_2023',
    'CUENTA_202': 'PASIVO_NO_CORRIENTE_2023',
    'CUENTA_3': 'PATRIMONIO_NETO_2023',
    'CUENTA_401': 'INGRESOS_ACTIVIDADES_ORDINARIAS_2023',
    'CUENTA_402': 'GANANCIA_BRUTA_2023',
    'CUENTA_403': 'OTROS_INGRESOS_2023',
    'CUENTA_501': 'COSTO_VENTAS_PRODUCCION_2023',
    'CUENTA_502': 'GASTOS_2023',
    'CUENTA_600': 'UTILIDAD_OPERATIVA_2023',
    'CUENTA_603': 'IMPUESTO_RENTA_2023',
    'CUENTA_602': 'GANACIA_PERDIDA_ANTES_IR_2023',
    'CUENTA_607': 'UTILIDAD_NETA_2023'
    }

pd_dataset_filtered = pd_dataset_filtered.rename(columns=diccionario_2023)

pd_dataset_filtered.sample(1)

pd_dataset_filtered.dtypes

### EDA

description = pd_dataset_filtered[['ACTIVO_2023', 'PASIVO_2023', 'PATRIMONIO_NETO_2023',
                                   'INGRESOS_ACTIVIDADES_ORDINARIAS_2023', 
                                   'COSTO_VENTAS_PRODUCCION_2023', 'GASTOS_2023',
                                   'UTILIDAD_OPERATIVA_2023', 'GANACIA_PERDIDA_ANTES_IR_2023', 'IMPUESTO_RENTA_2023', 'UTILIDAD_NETA_2023']].describe()

# Aplicar formato a los valores en el DataFrame de descripción
formatted_description = description.applymap(lambda x: f"{x:.2f}")
formatted_description

### Revisión casos puntuales 2023

pd_dataset_if_2023 = pd_dataset_filtered.copy()

# 14272  1792701341001  90.262597  -13.776378
# 26976  0993261068001  31.702556  219.933084
# 29587  0993372222001  16.300589   41.626631
# 9801   0992735120001  11.958218   82.614101
# 18806  0993261068001  28.606353  181.455280
# 20552  0993372222001  14.541973   34.095984
# 7313   0992735120001  10.800392   68.133014

pd_dataset_if_2023[pd_dataset_if_2023['RUC']=='1391932908001']

### Exportar Dataset 2023

print(f"Dimensiones final del conjunto de datos: \n Filas: {pd_dataset_if_2023.shape[0]}  \n Columnas: {pd_dataset_if_2023.shape[1]}")

pd_dataset_if_2023.to_csv('Datasets Procesados/informacion_financiera_2023.csv', index=False)


## PROCESAMIENTO Información Financiera 2022

### Cargar y Procesamiento Dataset

file_path = "Datasets\\estadosFinancieros_2022\\balances_2022_1.txt"
pd_dataset = pd.read_csv(file_path, dtype={'RUC': 'object', 'EXPEDIENTE': 'object'}, delimiter='\t', encoding='latin1')

pd_dataset_filtered = pd_dataset[['EXPEDIENTE', 'RUC', 
                                  'CUENTA_1', 'CUENTA_101', 'CUENTA_10103', 'CUENTA_102',
                                  'CUENTA_2', 'CUENTA_201', 'CUENTA_202', 
                                  'CUENTA_3', 
                                  'CUENTA_401', 'CUENTA_402', 'CUENTA_403', 
                                  'CUENTA_501','CUENTA_502', 
                                  'CUENTA_600', 'CUENTA_602', 'CUENTA_603', 'CUENTA_607' ]]

num_columns = ['CUENTA_1', 'CUENTA_101', 'CUENTA_10103', 'CUENTA_102', 
               'CUENTA_2', 'CUENTA_201', 'CUENTA_202', 
               'CUENTA_3', 
                'CUENTA_401', 'CUENTA_402', 'CUENTA_403', 
                'CUENTA_501', 'CUENTA_502', 
                'CUENTA_600','CUENTA_602', 'CUENTA_603', 'CUENTA_607' ]

pd_dataset_filtered[num_columns] = pd_dataset_filtered[num_columns].apply(
    lambda col: pd.to_numeric(col.str.replace('.', '', regex=False).str.replace(',', '.', regex=False))
)

diccionario_2022 = {
    'CUENTA_1': 'ACTIVO_2022',
    'CUENTA_101': 'ACTIVO_CORRIENTE_2022',
    'CUENTA_10103': 'INVENTARIOS_2022',
    'CUENTA_102': 'ACTIVOS_NO_CORRIENTES_2022',
    'CUENTA_2': 'PASIVO_2022',
    'CUENTA_201': 'PASIVO_CORRIENTE_2022',
    'CUENTA_202': 'PASIVO_NO_CORRIENTE_2022',
    'CUENTA_3': 'PATRIMONIO_NETO_2022',
    'CUENTA_401': 'INGRESOS_ACTIVIDADES_ORDINARIAS_2022',
    'CUENTA_402': 'GANANCIA_BRUTA_2022',
    'CUENTA_403': 'OTROS_INGRESOS_2022',
    'CUENTA_501': 'COSTO_VENTAS_PRODUCCION_2022',
    'CUENTA_502': 'GASTOS_2022',    
    'CUENTA_600': 'UTILIDAD_OPERATIVA_2022',
    'CUENTA_602': 'GANACIA_PERDIDA_ANTES_IR_2022',
    'CUENTA_603': 'IMPUESTO_RENTA_2022',
    'CUENTA_607': 'UTILIDAD_NETA_2022'
    }

pd_dataset_filtered = pd_dataset_filtered.rename(columns=diccionario_2022)

pd_dataset_if_2022 = pd_dataset_filtered.copy()

### Exportar Dataset 2022

print(f"Dimensiones final del conjunto de datos: \n Filas: {pd_dataset_if_2022.shape[0]}  \n Columnas: {pd_dataset_if_2022.shape[1]}")
pd_dataset_if_2022.to_csv('Datasets Procesados/informacion_financiera_2022.csv', index=False)

### Revisión de casos particulares 2022

# 14272  1792701341001  90.262597  -13.776378
# 26976  0993261068001  31.702556  219.933084
# 29587  0993372222001  16.300589   41.626631
# 9801   0992735120001  11.958218   82.614101
# 18806  0993261068001  28.606353  181.455280
# 20552  0993372222001  14.541973   34.095984
# 7313   0992735120001  10.800392   68.133014

pd_dataset_if_2022[pd_dataset_if_2022['RUC']=='1391932908001']
# Se detectan empresas que registran INGRESOS pero no EGRESOS ni impuesto a la renta.
