==> Notebooks/3_Compilacion_Dataset_Final_EDA.ipynb <==
# 3 - UNIFICAR DATSETS y EDA

**AUTOR: Fabrizio Ramirez Cutimbo**

**OBJETIVO**

+ Unificar todos los datasests de la Superintendencia de Compañías (directorio, información financiera y ciiu).
+ Aplicar técnicas de anonimización a datos identificadores como EXPEDIENTE y RUC para evitar conflicto de interés.
---

import pandas as pd
import numpy as np

### Carga de datasets

pd_companias = pd.read_csv('Datasets Procesados/directorio_empresas.csv', dtype={'RUC': 'object'})
pd_info_financiera_2022 = pd.read_csv('Datasets Procesados/informacion_financiera_2022.csv', dtype={'RUC': 'object'})
pd_info_financiera_2023 = pd.read_csv('Datasets Procesados/informacion_financiera_2023.csv', dtype={'RUC': 'object'})

### Forma original

print(f"Dimensión Original de directorio de Compañías. \n Filas: {pd_companias.shape[0]} \n Columnas: {pd_companias.shape[1]}")
pd_companias.sample(1)

print(f"Dimensión Original de datset Información Financiera 2022.\n Filas: {pd_info_financiera_2022.shape[0]} \n Columnas: {pd_info_financiera_2022.shape[1]}" )
pd_info_financiera_2022.sample(1)


print(f"Dimensión Original de datset Información Financiera 2023.\n Filas: {pd_info_financiera_2023.shape[0]} \n Columnas: {pd_info_financiera_2023.shape[1]}" )
pd_info_financiera_2023.sample(1)

## Combinación de datasets

print(f"Tamaño inicial de dataset compañias: {pd_companias.shape[0]}")

+ Se aplica un LEFT MERGE para conversar todas las compañías y unirlas con la información financiera de 2022 y 2023.

pd_merged = pd.merge(pd_companias, pd_info_financiera_2022, left_on='RUC', right_on='RUC', how='left')
pd_merged = pd.merge(pd_merged, pd_info_financiera_2023, left_on='RUC', right_on='RUC', how='left')

+ Se revisan los NAN en las columnas del año 2022 y 2023

# Se valida respecto a la columna ACTIVO como referencia
pd_validation_2022 = pd_merged[pd_merged['ACTIVO_2022'].isna()]
pd_validation_2023 = pd_merged[pd_merged['ACTIVO_2023'].isna()]
pd_validation_both = pd_merged[pd_merged[['ACTIVO_2022', 'ACTIVO_2023']].isna().all(axis=1)]

print(f'Registros con valonres NaN en Información Financiera 2022: {pd_validation_2022.shape[0]}')
print(f'Registros con valonres NaN en Información Financiera 2023: {pd_validation_2023.shape[0]}')
print(f'Registros con valonres NaN en Información Financiera 2022 y 2023: {pd_validation_both.shape[0]}')

+ Si no existe un valor en al columna 2022 se reemplaza el correspondiente del año 2023 y viceversa

# Lista base de columnas a revisar
columnas_reemplazo = [
    'ACTIVO', 'ACTIVO_CORRIENTE', 'INVENTARIOS', 'ACTIVOS_NO_CORRIENTES', 
    'PASIVO', 'PASIVO_CORRIENTE', 'PASIVO_NO_CORRIENTE', 
    'PATRIMONIO_NETO',
    'INGRESOS_ACTIVIDADES_ORDINARIAS', 'GANANCIA_BRUTA', 'OTROS_INGRESOS',
    'COSTO_VENTAS_PRODUCCION', 'GASTOS', 
    'UTILIDAD_OPERATIVA', 'GANACIA_PERDIDA_ANTES_IR', 'IMPUESTO_RENTA', 'UTILIDAD_NETA'
]

# Procesar cada columna base
for columna in columnas_reemplazo:
    col_anio1 = f"{columna}_2022"
    col_anio2 = f"{columna}_2023"
    
    # Verificar que ambas columnas existan
    if col_anio1 in pd_merged.columns and col_anio2 in pd_merged.columns:

        # Rellenar NaN del año1 con los valores del año2
        pd_merged[col_anio1] = pd_merged[col_anio1].fillna(pd_merged[col_anio2])
        # Rellenar NaN del año2 con los valores del año1
        pd_merged[col_anio2] = pd_merged[col_anio2].fillna(pd_merged[col_anio1])

pd_validation_2022 = pd_merged[pd_merged['ACTIVO_2022'].isna()]
pd_validation_2023 = pd_merged[pd_merged['ACTIVO_2023'].isna()]
pd_validation_both = pd_merged[pd_merged[['ACTIVO_2022', 'ACTIVO_2023']].isna().all(axis=1)]

print(f'Registros con valonres NaN en Información Financiera 2022: {pd_validation_2022.shape[0]}')
print(f'Registros con valonres NaN en Información Financiera 2023: {pd_validation_2023.shape[0]}')
print(f'Registros con valonres NaN en Información Financiera 2022 y 2023: {pd_validation_both.shape[0]}')

+ Si no existe un valor en la columna del 2022 y 2023 se elimina la fila

# Filtrar filas donde las columnas de 2022 y 2023 tengan datos
pd_merged_cleaned = pd_merged.dropna(subset=['ACTIVO_2022', 'ACTIVO_2023'], how='all')
print(f"Tamaño luego de combinación y limpieza: {pd_merged_cleaned.shape[0]}")


- Estadisticas del dataset antes de filtros

# función para imprimir Describe() con dos decimales
def describe_formateado(dataset, columns):

    # Selección de las columnas relevantes
    description = dataset[columns].describe()
    # Aplicar formato a los valores en el DataFrame de descripción
    formatted_description = description.applymap(lambda x: f"{x:.3f}")
    return formatted_description

describe_formateado(pd_merged_cleaned, ['ACTIVO_2023', 'ACTIVO_CORRIENTE_2023', 'INVENTARIOS_2023', 'ACTIVOS_NO_CORRIENTES_2023', 
                                        'PASIVO_2023', 'PASIVO_CORRIENTE_2023', 'PASIVO_NO_CORRIENTE_2023', 'PATRIMONIO_NETO_2023', 
                                        'INGRESOS_ACTIVIDADES_ORDINARIAS_2023', 'GANANCIA_BRUTA_2023', 'OTROS_INGRESOS_2023', 
                                        'COSTO_VENTAS_PRODUCCION_2023', 'GASTOS_2023', 
                                        'UTILIDAD_OPERATIVA_2023', 'GANACIA_PERDIDA_ANTES_IR_2023', 'IMPUESTO_RENTA_2023', 'UTILIDAD_NETA_2023'])

**IMPORTANTE** Se puede observar que varias variables como los "INVENTARIOS", "IMPUESTO_RENTA_2023".
Recien en el tercer Cuartil 3 tiene un valor diferente a cero, esto ocurre porque la gran mayoría de valores tienen el valor 0

## Limpieza casos anómalos
+ Se eliminan empresas que no hayan registrado ningun tipo de ingreso en 2023  y 2022.
+ Los **INGRESOS** son un dato fundamental pues se utilizarán en el Location Quotient.

condicion = (
    (pd_merged_cleaned['INGRESOS_ACTIVIDADES_ORDINARIAS_2022'] == 0) &
    (pd_merged_cleaned['GANANCIA_BRUTA_2022'] == 0) &
    (pd_merged_cleaned['OTROS_INGRESOS_2022'] == 0) &
    (pd_merged_cleaned['INGRESOS_ACTIVIDADES_ORDINARIAS_2023'] == 0) &
    (pd_merged_cleaned['GANANCIA_BRUTA_2023'] == 0) &
    (pd_merged_cleaned['OTROS_INGRESOS_2023'] == 0)
)

# Conteo registros que cumplen con la condición
pd_empresas_sin_ingresos = pd_merged_cleaned[condicion]
registros_cumplen_condicion = pd_merged_cleaned[condicion].shape[0]
print(f"Cantidad de Empresas ACTIVAS que NO registran INGRESOS: {registros_cumplen_condicion}")

print(f"% De Empresas que no registran ingresos: ")
round(registros_cumplen_condicion/pd_merged_cleaned.shape[0]*100, 2)

print(pd_empresas_sin_ingresos.shape[0])
pd_empresas_sin_ingresos.sample(1)

describe_formateado(pd_empresas_sin_ingresos, ['ACTIVO_2022', 'ACTIVO_2023', 'PASIVO_2022', 'PASIVO_2023', 
                                               'INGRESOS_ACTIVIDADES_ORDINARIAS_2022', 'INGRESOS_ACTIVIDADES_ORDINARIAS_2023',                                               
                                               'COSTO_VENTAS_PRODUCCION_2022', 'COSTO_VENTAS_PRODUCCION_2023', 'GASTOS_2022', 'GASTOS_2023',
                                               'UTILIDAD_OPERATIVA_2022', 'UTILIDAD_OPERATIVA_2023'])

**Importante** Se observan empresas que durante dos años consecutivos no han generado ningun tipo de ingreso por actividades ordinarias ni extraordinarias, la gran mayoría tampoco tiene costos ni gastos. Potencial revisión para trabajos futuros.

+ Eliminar las filas que cumplen con la condición

**IMPORTANTE**, aquí se observa que se pierde un conjunto considerable de Empresas

print(f'Tamaño antes del filtro: {pd_merged_cleaned.shape[0]}')

# Eliminar las filas que cumplen con la condición
pd_merged_cleaned = pd_merged_cleaned[~condicion]

print(f"Tamaño después de filtro: {pd_merged_cleaned.shape[0]}")

+ Eliminar empresas cuyo impuesto a la renta sea mayor o igual que los Ingresos (lo cual es extraño pues el IR es una fracción de los ingresos)

condicion = (
    (pd_merged_cleaned['IMPUESTO_RENTA_2023'] > (1.0 * (
        pd_merged_cleaned['INGRESOS_ACTIVIDADES_ORDINARIAS_2023'] +
        pd_merged_cleaned['GANANCIA_BRUTA_2023'] +
        pd_merged_cleaned['OTROS_INGRESOS_2023']
    ))) &
    (pd_merged_cleaned['UTILIDAD_OPERATIVA_2023'] > 0)
)

# Conteo registros que cumplen con la condición
registros_cumplen_condicion = pd_merged_cleaned[condicion].shape[0]
print(f"Cantidad de registros que cumplen la condición: {registros_cumplen_condicion}")

print(f'Tamaño antes del filtro: {pd_merged_cleaned.shape[0]}')

# Eliminar las filas que cumplen con la condición
pd_merged_cleaned = pd_merged_cleaned[~condicion]

print(f"Tamaño después de filtro: {pd_merged_cleaned.shape[0]}")

### Seleccionar y Ordenar Columnas

pd_merged_cleaned.columns

pd_dataset_final = pd_merged_cleaned[['EXPEDIENTE_x', 'RUC', 'SITUACION_LEGAL', 'FECHA_CONSTITUCION', 'TIPO', 'PAIS', 'REGION', 'PROVINCIA', 'CANTON', 'CIUDAD', 
       'CIIU_NIVEL_1', 'CIIU_NIVEL_1_DESC', 'CIIU_NIVEL_3', 'CIIU_NIVEL_3_DESC', 'CIIU_NIVEL_4', 'CIIU_NIVEL_4_DESC', 'CIIU_NIVEL_6', 'CIIU_NIVEL_6_DESC',
       # Información Financiera 2022
       'ACTIVO_2022', 'ACTIVO_CORRIENTE_2022', 'INVENTARIOS_2022', 'ACTIVOS_NO_CORRIENTES_2022', 
       'PASIVO_2022', 'PASIVO_CORRIENTE_2022', 'PASIVO_NO_CORRIENTE_2022', 'PATRIMONIO_NETO_2022',
       'INGRESOS_ACTIVIDADES_ORDINARIAS_2022', 'GANANCIA_BRUTA_2022', 'OTROS_INGRESOS_2022', 
       'COSTO_VENTAS_PRODUCCION_2022', 'GASTOS_2022', 
       'UTILIDAD_OPERATIVA_2022', 'GANACIA_PERDIDA_ANTES_IR_2022', 'IMPUESTO_RENTA_2022', 'UTILIDAD_NETA_2022', 
       # Información Financiera 2023
       'ACTIVO_2023', 'ACTIVO_CORRIENTE_2023','INVENTARIOS_2023', 'ACTIVOS_NO_CORRIENTES_2023', 
       'PASIVO_2023', 'PASIVO_CORRIENTE_2023', 'PASIVO_NO_CORRIENTE_2023', 'PATRIMONIO_NETO_2023', 
       'INGRESOS_ACTIVIDADES_ORDINARIAS_2023', 'GANANCIA_BRUTA_2023', 'OTROS_INGRESOS_2023', 
       'COSTO_VENTAS_PRODUCCION_2023', 'GASTOS_2023', 
       'UTILIDAD_OPERATIVA_2023', 'GANACIA_PERDIDA_ANTES_IR_2023', 'IMPUESTO_RENTA_2023', 'UTILIDAD_NETA_2023']]

# Renombrar columna
pd_dataset_final = pd_dataset_final.rename(columns={
    'EXPEDIENTE_x':'EXPEDIENTE',
    'NOMBRE_x':'NOMBRE'
})
print(f"Tamaño del datset: {pd_dataset_final.shape[0]}")

## Ajustar valores pequeños

- Se revisan las columnas que se utilizarán para calcular variaciones e indicadores financieros.
- Dado que las variaciones e indicadores financieros realizan **divisiones**, además de controlar la división para 0, también se debe validar que no existan valores muy pequeños que puedan incrementar considerablemente el numerador.
- Si se detectan valores entre -1 y 1, excluyendo el 0, este valor se reemplazará redondeará.

# Columnas a Revisar
columns_revision = ['ACTIVO_2022', 'ACTIVO_CORRIENTE_2022', 'INVENTARIOS_2022', 'ACTIVOS_NO_CORRIENTES_2022', 
                    'PASIVO_2022', 'PASIVO_CORRIENTE_2022', 'PASIVO_NO_CORRIENTE_2022', 'PATRIMONIO_NETO_2022',
                    'INGRESOS_ACTIVIDADES_ORDINARIAS_2022', 'GANANCIA_BRUTA_2022', 'OTROS_INGRESOS_2022', 
                    'COSTO_VENTAS_PRODUCCION_2022', 'GASTOS_2022', 
                    'UTILIDAD_OPERATIVA_2022', 'GANACIA_PERDIDA_ANTES_IR_2022', 'IMPUESTO_RENTA_2022', 'UTILIDAD_NETA_2022', 
                    'ACTIVO_2023', 'ACTIVO_CORRIENTE_2023','INVENTARIOS_2023', 'ACTIVOS_NO_CORRIENTES_2023', 
                    'PASIVO_2023', 'PASIVO_CORRIENTE_2023', 'PASIVO_NO_CORRIENTE_2023', 'PATRIMONIO_NETO_2023', 
                    'INGRESOS_ACTIVIDADES_ORDINARIAS_2023', 'GANANCIA_BRUTA_2023', 'OTROS_INGRESOS_2023', 
                    'COSTO_VENTAS_PRODUCCION_2023', 'GASTOS_2023', 
                    'UTILIDAD_OPERATIVA_2023', 'GANACIA_PERDIDA_ANTES_IR_2023', 'IMPUESTO_RENTA_2023', 'UTILIDAD_NETA_2023']

# Crear una máscara booleana para valores entre -1 y 1, excluyendo el 0
pd_dataset_anomalos = pd_dataset_final[columns_revision].apply(
    lambda col: col.map(lambda x: -1 < x < 1 and x != 0)
)

# Filtrar las filas donde al menos una columna cumpla la condición
pd_dataset_anomalos = pd_dataset_final[pd_dataset_anomalos.any(axis=1)]

# Mostrar los registros encontrados
print(f"Registros anómalos con valores entre -1 y -1 (Excluyendo 0): {pd_dataset_anomalos.shape[0]}")

pd_dataset_anomalos['OTROS_INGRESOS_2023'].head(3)

+ Redondear números anómalos entre -1 y 1

for col in columns_revision:
    # Aplicamos el redondeo solo a los valores entre -1 y 1, excluyendo el 0
    pd_dataset_final[col] = pd_dataset_final[col].apply(lambda x: round(x) if -1 < x < 1 and x != 0 else x)


# Mostrar el dataset actualizado
pd_dataset_final[columns_revision].head(3)

describe_formateado(pd_dataset_final, columns_revision)

print(f"Tamaño del datset luego de Filtros: {pd_dataset_final.shape[0]}")

## Enriquecimiento: Crear nuevas columnas

### Variación de Cuentas Financieras

+ **Variación de Cuentas Financieras** respecto al año anterior
+ Considerar que hay empresas que tienen cuentas en 0, incluso Activos. Para prevenir divisiones para 0, si el divisor es 0, entonces se establece automáticamente "0", que en teoría indica no hay variación: (10 - 10) / 10 = 0

pd_dataset_final['VAR_ACTIVO'] = pd_dataset_final.apply(lambda row: round(((row['ACTIVO_2023'] - row['ACTIVO_2022']) / row['ACTIVO_2022']) if row['ACTIVO_2022'] != 0 else 0, 3), axis=1)
pd_dataset_final['VAR_INVENTARIOS'] = pd_dataset_final.apply(lambda row: round(((row['INVENTARIOS_2023'] - row['INVENTARIOS_2022']) / row['INVENTARIOS_2022']) if row['INVENTARIOS_2022'] != 0 else 0, 3), axis=1)
pd_dataset_final['VAR_PASIVO'] = pd_dataset_final.apply(lambda row: round(((row['PASIVO_2023'] - row['PASIVO_2022']) / row['PASIVO_2022']) if row['PASIVO_2022'] != 0 else 0, 3), axis=1)
pd_dataset_final['VAR_INGRESOS'] = pd_dataset_final.apply(lambda row: round(((row['INGRESOS_ACTIVIDADES_ORDINARIAS_2023'] - row['INGRESOS_ACTIVIDADES_ORDINARIAS_2022']) / row['INGRESOS_ACTIVIDADES_ORDINARIAS_2022']) if row['INGRESOS_ACTIVIDADES_ORDINARIAS_2022'] != 0 else 0, 3), axis=1)
pd_dataset_final['VAR_COSTO_VENTAS_PRODUCCION'] = pd_dataset_final.apply(lambda row: round(((row['COSTO_VENTAS_PRODUCCION_2023'] - row['COSTO_VENTAS_PRODUCCION_2022']) / row['COSTO_VENTAS_PRODUCCION_2022']) if row['COSTO_VENTAS_PRODUCCION_2022'] != 0 else 0, 3), axis=1)
pd_dataset_final['VAR_GASTOS'] = pd_dataset_final.apply(lambda row: round(((row['GASTOS_2023'] - row['GASTOS_2022']) / row['GASTOS_2022']) if row['GASTOS_2022'] != 0 else 0, 3), axis=1)
pd_dataset_final['VAR_UTILIDAD_OPERATIVA'] = pd_dataset_final.apply(lambda row: round(((row['UTILIDAD_OPERATIVA_2023'] - row['UTILIDAD_OPERATIVA_2022']) / row['UTILIDAD_OPERATIVA_2022']) if row['UTILIDAD_OPERATIVA_2022'] != 0 else 0, 3), axis=1)
pd_dataset_final['VAR_IMPUESTO_RENTA'] = pd_dataset_final.apply(lambda row: round(((row['IMPUESTO_RENTA_2023'] - row['IMPUESTO_RENTA_2022']) / row['IMPUESTO_RENTA_2022']) if row['IMPUESTO_RENTA_2022'] != 0 else 0, 3), axis=1)
pd_dataset_final['VAR_UTILIDAD_NETA'] = pd_dataset_final.apply(lambda row: round(((row['UTILIDAD_NETA_2023'] - row['UTILIDAD_NETA_2022']) / row['UTILIDAD_NETA_2022']) if row['UTILIDAD_NETA_2022'] != 0 else 0, 3), axis=1)


# Visualizar una muestra
pd_dataset_final[['RUC','ACTIVO_2022','ACTIVO_2023','VAR_ACTIVO']].sample(3)

+ Visualizar Estadísticas de la variaciones

**NOTA** Se observan valores altos de desviación estándar

columnas_variacion = ['VAR_ACTIVO', 'VAR_INVENTARIOS', 'VAR_PASIVO', 'VAR_INGRESOS','VAR_COSTO_VENTAS_PRODUCCION', 'VAR_GASTOS', 'VAR_UTILIDAD_OPERATIVA', 'VAR_IMPUESTO_RENTA', 'VAR_UTILIDAD_NETA']
describe_formateado(pd_dataset_final, columnas_variacion)

+ Importante notar que 'VAR_INVENTARIOS', 'VAR_COSTO_VENTAS_PRODUCCION', 'VAR_IMPUESTO_RENTA' en los Q1, Q2 y Q3 tiene valor en 0, lo cual indica que la gran mayoria de valores son "0"

print("Tamaño Dataset")
print(pd_dataset_final.shape)

Aplicar **Rango Inter Cuartílico** (Extendido hasta *10 el Q1 y Q3)

Se utiliza 10 para ser más flexibles y solo excluir valores extremos

# Función para control de Outliers, no se eliminan sino se reemplazan por los limites definidos
def eliminar_outliers_tecnica_iqr(dataset, columns, iqr_multiplier=1.5, lower_quantile=0.25, upper_quantile=0.75):

    cleaned_df = dataset.copy()  # Crear una copia del DataFrame original

    for col in columns:
        # Calcular Q1, Q3 e IQR
        Q1 = cleaned_df[col].quantile(lower_quantile) # Por defecto 25%
        Q3 = cleaned_df[col].quantile(upper_quantile) # Por defecto 75%
        IQR = Q3 - Q1

        # Definir límites inferior y superior
        lower_bound = Q1 - iqr_multiplier * IQR
        upper_bound = Q3 + iqr_multiplier * IQR

        # Reemplazar valores menores que el límite inferior
        cleaned_df.loc[cleaned_df[col] < lower_bound, col] = lower_bound

        # Reemplazar valores mayores que el límite superior
        cleaned_df.loc[cleaned_df[col] > upper_bound, col] = upper_bound

    return cleaned_df

columns_to_clean = ['VAR_ACTIVO', 'VAR_PASIVO', 
                    'VAR_INGRESOS', 'VAR_GASTOS', 
                    'VAR_UTILIDAD_OPERATIVA', 'VAR_UTILIDAD_NETA']

columns_to_clean_2 = ['VAR_INVENTARIOS', 'VAR_COSTO_VENTAS_PRODUCCION', 'VAR_IMPUESTO_RENTA']

pd_dataset_final = eliminar_outliers_tecnica_iqr(pd_dataset_final, columns_to_clean, 10)

# En Las columnas 'VAR_INVENTARIOS', 'VAR_IMPUESTO_RENTA' se realiza un proceso más flexible del 10% y 90%
pd_dataset_final = eliminar_outliers_tecnica_iqr(pd_dataset_final, columns_to_clean_2, 10, 0.1, 0.9)

columnas = ['VAR_ACTIVO', 'VAR_INVENTARIOS', 'VAR_PASIVO', 'VAR_INGRESOS','VAR_COSTO_VENTAS_PRODUCCION', 'VAR_GASTOS', 'VAR_UTILIDAD_OPERATIVA', 'VAR_IMPUESTO_RENTA', 'VAR_UTILIDAD_NETA']
describe_formateado(pd_dataset_final, columnas)

+ Importante Notar como se reduce la Desviación Estándar

### Segmento de la Compañía

Se crea una nueva columna para segmentar las empresas de acuerdo con la ley ecuatoriana de segmentación, que es una buena referencia para determinar la dimensión de una empresa:
+ MIPE
+ PYME
+ EMPRESARIAL
+ CORPORARTIVO

def segmento_empresarial(ingresos):
    if ingresos < 100000:
        return 'MICRO'
    elif 100000 <= ingresos < 1000000:
        return 'PEQUENA'
    elif 1000000 <= ingresos < 5000000:
        return 'MEDIANA'
    elif ingresos >= 5000000:
        return 'GRANDE'
    else:
        return 'OTRO'

# Se aplica segmentación en función de los ingresos de Actividades Ordinarias
pd_dataset_final['SEGMENTO'] = pd_dataset_final['INGRESOS_ACTIVIDADES_ORDINARIAS_2023'].apply(segmento_empresarial)


pd_dataset_final[['INGRESOS_ACTIVIDADES_ORDINARIAS_2023', 'SEGMENTO']].sample(4)

- Resumen de empresas agrupadas por segmento

pd_dataset_final['SEGMENTO'].value_counts().sum

### Indicadores Financieros

| RAZÓN         | INDICADORES                  | FÓRMULA                                           |
|---------------|------------------------------|--------------------------------------------------|
| **Liquidez**  | Liquidez corriente           | Activo Corriente / Pasivo Corriente             |
|               | Prueba ácida                 | (Activo Corriente – Inventarios) / Pasivo Corriente |
| **Solvencia** | Endeudamiento del activo     | Pasivo Total / Activo Total                     |
|               | Endeudamiento patrimonial    | Pasivo Total / Patrimonio                       |
|               | Endeudamiento del Activo fijo| Patrimonio / Activo Fijo                        |
|               | Endeudamiento a corto plazo  | Pasivo Corriente / Pasivo Total                 |
|               | Endeudamiento a largo plazo  | Pasivo No Corriente / Pasivo Total              |
| **Gestión**   | Rotación de ventas           | Ventas / Activo Total                           |
| **Rentabilidad** | Margen Bruto              | (Ventas - Costo De Ventas) / Ventas             |
|               | Margen Operacional           | Utilidad Operacional / Ventas                   |
|               | ROE                          | Utilidad Neta / Patrimonio                      |
|               | ROA                          | Utilidad Neta / Activo Total                    |

* Utilidad Neta: Después del 15% de trabajadores e impuesto a la renta (Equivale a la "Cuenta 607")


**Nota** Si en el cálculo de algún Indicador Financiero resulta en división para 0. Entonces se establecerá el valor 1, lo cual en la mayoría de casos indica un valor neutro en la mayoría de interpretaciones.

Este tipo de formulas, son susceptibles a outliers, pues si una empresa registra 1 dolar de activos, patrimonio, pasivo, etc. Las divisiones generarán número muy grandes fuera de los rangos interpretativos normales. Será necesario aplicar técnicas de control de outliers luego del cálculo

Como trabajo futuro se podría, establecer estos casos como NaN y especificar el valor promedio del sector

# LIQUIDEZ
pd_dataset_final['IF_PRUEBA_ACIDA'] = pd_dataset_final.apply(lambda row: round(((row['ACTIVO_CORRIENTE_2023'] - row['INVENTARIOS_2023']) / row['PASIVO_CORRIENTE_2023']) if row['PASIVO_CORRIENTE_2023'] != 0 else 1, 3), axis=1)

#  SOLVENCIA
pd_dataset_final['IF_ENDEUDAMIENTO_ACTIVO'] = pd_dataset_final.apply(lambda row: round((row['PASIVO_2023'] / row['ACTIVO_2023']) if row['ACTIVO_2023'] != 0 else 1, 3), axis=1)
pd_dataset_final['IF_APALANCAMIENTO'] = pd_dataset_final.apply(lambda row: round((row['ACTIVO_2023'] / row['PATRIMONIO_NETO_2023']) if row['PATRIMONIO_NETO_2023'] != 0 else 1, 3), axis=1)

# GESTION
pd_dataset_final['IF_ROTACION_VENTAS'] = pd_dataset_final.apply(lambda row: round((row['INGRESOS_ACTIVIDADES_ORDINARIAS_2023'] / row['ACTIVO_2023']) if row['ACTIVO_2023'] != 0 else 1, 3), axis=1)

# RENTABILIDAD
pd_dataset_final['IF_MARGEN_BRUTO'] = pd_dataset_final.apply(lambda row: round(((row['INGRESOS_ACTIVIDADES_ORDINARIAS_2023'] - row['COSTO_VENTAS_PRODUCCION_2023']) / row['INGRESOS_ACTIVIDADES_ORDINARIAS_2023']) if row['INGRESOS_ACTIVIDADES_ORDINARIAS_2023'] != 0 else 1, 3), axis=1)
pd_dataset_final['IF_MARGEN_OPERACIONAL'] = pd_dataset_final.apply(lambda row: round(((row['UTILIDAD_OPERATIVA_2023']) / row['INGRESOS_ACTIVIDADES_ORDINARIAS_2023']) if row['INGRESOS_ACTIVIDADES_ORDINARIAS_2023'] != 0 else 1, 3), axis=1)
pd_dataset_final['IF_MARGEN_NETO'] = pd_dataset_final.apply(lambda row: round(((row['UTILIDAD_NETA_2023']) / row['INGRESOS_ACTIVIDADES_ORDINARIAS_2023']) if row['INGRESOS_ACTIVIDADES_ORDINARIAS_2023'] != 0 else 1, 3), axis=1)
pd_dataset_final['IF_ROE'] = pd_dataset_final.apply(lambda row: round((row['UTILIDAD_NETA_2023'] / row['PATRIMONIO_NETO_2023']) if row['PATRIMONIO_NETO_2023'] != 0 else 1, 3), axis=1)
pd_dataset_final['IF_ROA'] = pd_dataset_final.apply(lambda row: round((row['UTILIDAD_NETA_2023'] / row['ACTIVO_2023']) if row['ACTIVO_2023'] != 0 else 1, 3), axis=1)


columns_to_clean_if = ['IF_PRUEBA_ACIDA','IF_ENDEUDAMIENTO_ACTIVO','IF_APALANCAMIENTO','IF_ROTACION_VENTAS','IF_MARGEN_BRUTO','IF_MARGEN_OPERACIONAL','IF_MARGEN_NETO','IF_ROE','IF_ROA']
describe_formateado(pd_dataset_final, columns_to_clean_if)


Aplicar control de outlier Extremos

## Aplicar control de outlier Extremos
pd_dataset_final = eliminar_outliers_tecnica_iqr(pd_dataset_final, columns_to_clean_if, 10)


#Describe luego de limpieza
describe_formateado(pd_dataset_final, columns_to_clean_if)

## Exportar DATASET FINAL COMPLETO para EDA
Se exporta el datset completo, conservando todas las actividades Económicas, para:
+ Realizar un EDA 
+ Calcular el Location Quotient

print(f'Tamaño Dataset: {pd_dataset_final.shape}')
pd_dataset_final.to_csv('Datasets Procesados\\directorio_empresas_final_ciiu_completo.csv', index=False)

## Exportar DATSET FINAL Para CLUSTERING

Se consideran las siguientes categorías CIIU (SECTORES)
+ A - AGRICULTURA, GANADERÍA, SILVICULTURA Y PESCA
+ C - INDUSTRIAS MANUFACTURERAS
+ I - ALOJAMIENTO Y COMIDAS (Servicios)
+ J - INFORMACIÓN Y COMUNICACIÓN. (Servicios)
+ K - ACTIVIDADES FINANCIERAS Y DE SEGUROS. (Servicios)
+ L - ACTIVIDADES INMOBILIARIAS. (Excluida tras primera iteración pruebas)
+ M - Actividades profesionales, científicas y técnicas (Servicios)
+ N - Servicios administrativos y de apoyo (Servicios)

|  | Sección | Descripción Grande                                      | Sector                  |
|----|---------|---------------------------------------------------------|-------------------------|
|✅| A       | Agricultura, ganadería, silvicultura y pesca           | Agricultura, ganadería, silvicultura y pesca |
|   | B       | Explotación de minas y canteras                        | Explotación de minas y canteras |
|✅| C       | Industrias manufactureras                              | Industrias manufactureras |
|   | G       | Comercio al por mayor y al por menor                   | Comercio               |
|   | F       | Construcción                                           | Construcción           |
|   | D       | Suministro de electricidad, gas, vapor                 | Servicios              |
|   | E       | Agua, alcantarillado, desechos y saneamiento           | Servicios              |
|   | H       | Transporte y almacenamiento                            | Servicios              |
|✅| I       | Alojamiento y de servicios de comidas                  | Servicios              |
|✅| J       | Información y comunicación                             | Servicios              |
|✅| K       | Actividades financieras y de seguros                   | Servicios              |
|   | L       | Actividades inmobiliarias                              | Servicios              |
|✅| M       | Actividades profesionales, científicas y técnicas      | Servicios              |
|✅| N       | Servicios administrativos y de apoyo                   | Servicios              |
|   | O       | Administración pública, defensa, seguridad social      | Servicios              |
|   | P       | Enseñanza                                              | Servicios              |
|   | Q       | Atención de la salud humana y asistencia social        | Servicios              |
|   | R       | Artes, entretenimiento y recreación                    | Servicios              |
|   | S       | Otras actividades de servicios                         | Servicios              |



print('Dataset Shape Antes: ', pd_dataset_final.shape)
pd_dataset_final_filtrado = pd_dataset_final[pd_dataset_final['CIIU_NIVEL_1'].isin(['A', 'C', 'I', 'J', 'K', 'M', 'N'])]
print('Dataset Shape Despues: ', pd_dataset_final_filtrado.shape)

+ Para PCA se eliminan valores extremos en las cuentas de 2022 y 2023, luego de haber generado los Indicadores Financieros y variaciones

columns_revision = ['ACTIVO_2022', 'INVENTARIOS_2022', 'PASIVO_2022','PASIVO_CORRIENTE_2022', 
					'PATRIMONIO_NETO_2022',
                    'INGRESOS_ACTIVIDADES_ORDINARIAS_2022', 'COSTO_VENTAS_PRODUCCION_2022', 'GASTOS_2022',
                    'UTILIDAD_OPERATIVA_2022', 'GANACIA_PERDIDA_ANTES_IR_2022', 'IMPUESTO_RENTA_2022', 'UTILIDAD_NETA_2022',
                    'ACTIVO_2023', 'INVENTARIOS_2023', 'PASIVO_2023', 'PASIVO_CORRIENTE_2023', 
					'PATRIMONIO_NETO_2023',
                    'INGRESOS_ACTIVIDADES_ORDINARIAS_2023', 'COSTO_VENTAS_PRODUCCION_2023', 'GASTOS_2023',
                    'UTILIDAD_OPERATIVA_2023', 'GANACIA_PERDIDA_ANTES_IR_2023', 'IMPUESTO_RENTA_2023', 'UTILIDAD_NETA_2023']

pd_dataset_final_filtrado = eliminar_outliers_tecnica_iqr(pd_dataset_final_filtrado, columns_revision, 50)

describe_formateado(pd_dataset_final_filtrado, columns_revision)

print(f'Tamaño Dataset: {pd_dataset_final_filtrado.shape}')
pd_dataset_final_filtrado.to_csv('Datasets Procesados\\directorio_empresas_final.csv', index=False)
