==> Notebooks/8_Compilacion_Clusters.ipynb <==
# 8 - COMPILACIÓN CLUSTERS

**AUTOR: Fabrizio Ramirez Cutimbo**

**OBJETIVO**

+ Combinar en un solo dataset los clusters generados por PCA-Kmeans y UMAP-DBSCAN
---

import pandas as pd
import numpy as np
import hashlib
import json

### Cargar Datasets

## DATASET ORIGINAL COMPILADO
pd_companias = pd.read_csv('Datasets Procesados/directorio_empresas_final.csv', dtype={'RUC': 'object'})

## DATASET CON CLUSTERS
pd_companias_kmeans = pd.read_csv('Datasets Procesados/directorio_empresas_final_PCA_KMEANS.csv', dtype={'RUC': 'object'})
pd_companias_dbscan = pd.read_csv('Datasets Procesados/directorio_empresas_final_UMAP_DBSCAN.csv', dtype={'RUC': 'object'})

pd_companias.columns

print(pd_companias.shape)
pd_companias_kmeans.sample(1)

print(pd_companias_kmeans.shape)
pd_companias_kmeans.sample(1)

print(pd_companias_dbscan.shape)
pd_companias_dbscan.sample(1)

Se puede ver que todos los dataset contienen los mismos registros

pd_companias_kmeans.columns[pd_companias.shape[1]:]

pd_companias_dbscan.columns[pd_companias.shape[1]:]

## Combinar Datasets

Se selecciona únicamente los componentes principales

pd_merged = pd_companias

# pd_merged = pd.merge(pd_merged, pd_companias_kmeans[['RUC','Set10_PCA1', 'Set10_PCA2', 'Set10_REGION_PCA1', 'Set10_REGION_PCA2', 'CLUSTERS_KMEANS_SET_10', 'CLUSTERS_KMEANS_SET_10_REGION']], left_on='RUC', right_on='RUC', how='inner')
pd_merged = pd.merge(pd_merged, pd_companias_kmeans[['RUC','Set10_PCA1', 'Set10_PCA2', 'CLUSTERS_KMEANS_SET_10']], left_on='RUC', right_on='RUC', how='inner')
print(pd_merged.shape)

# pd_merged = pd.merge(pd_merged, pd_companias_dbscan[['RUC','Set4_UMAP1', 'Set4_UMAP2', 'Set4_REGION_UMAP1', 'Set4_REGION_UMAP2', 'CLUSTER_DBSCAN_SET4']], left_on='RUC', right_on='RUC', how='inner')
pd_merged = pd.merge(pd_merged, pd_companias_dbscan[['RUC','Set4_UMAP1', 'Set4_UMAP2', 'CLUSTER_DBSCAN_SET_4']], left_on='RUC', right_on='RUC', how='inner')
print(pd_merged.shape)

pd_merged.columns

### ANONIMIZACION 

# Función para aplicar hash
def hash_value(value):
    # Convertir el valor a string y aplicar hash
    return hashlib.new('sha256', str(value).encode()).hexdigest()

# Crear una copia del DataFrame y aplicar el hash
dataset_hashed = pd_merged.copy()
dataset_hashed['EXPEDIENTE'] = dataset_hashed['EXPEDIENTE'].apply(hash_value)

dataset_hashed.head(3)

# Eliminar las columnas
# dataset_hashed.drop(columns=['EXPEDIENTE', 'RUC', 'NOMBRE'], inplace=True)

dataset_hashed.drop(columns=['RUC'], inplace=True)


dataset_hashed.rename(columns={'EXPEDIENTE': 'ID'}, inplace=True)

dataset_hashed['CIIU_NIVEL_3_DESC'] = dataset_hashed['CIIU_NIVEL_3_DESC'].str.capitalize()
dataset_hashed['CIIU_NIVEL_4_DESC'] = dataset_hashed['CIIU_NIVEL_4_DESC'].str.capitalize()
dataset_hashed['CIIU_NIVEL_6_DESC'] = dataset_hashed['CIIU_NIVEL_6_DESC'].str.capitalize()
dataset_hashed['PROVINCIA'] = dataset_hashed['PROVINCIA'].str.capitalize()


# En la columna SEGMENTO, reemplazar PEQUENA por PEQUEÑA
dataset_hashed['SEGMENTO'] = dataset_hashed['SEGMENTO'].str.replace('PEQUENA', 'PEQUEÑA')

#### Añadir datos de Ubicación Geográfica

En Power BI no se puede identificara a la provincia de Bolivar, se confunde con Colombia y Venezuela
Se intentó de varias formas y no se pudo ubicar adecuadamente en el mapa. Se comenta este tipo de enriquecimiento.

# dataset_hashed['PROVINCIA'] = dataset_hashed['PROVINCIA'].replace('Bolivar', 'Bolívar')
# dataset_hashed['PROVINCIA_EC'] = dataset_hashed['PROVINCIA'] + ', Ecuador'

Se trabaja con Lattud y Longitud a nivel de Cantones Ecuador

# Se carga el dataset que tiene información de los cantones (división geográfica menor a provincia)
pd_cantones = pd.read_csv('Utilitarios/cantones_latitudes_longitudes.csv', encoding='utf-8')
pd_cantones.sample(5)

# Realizamos el MERGE entre los dos datasets por las columnas 'Provincia' y 'CIIU_NIVEL_3_DESC'
dataset_hashed = pd.merge(dataset_hashed,                       
                             pd_cantones,
                             on=['CANTON'], 
                             how='left')

dataset_hashed.shape

## Generar Dataset Compilado

Datset para caracterización en Power BI

# Guardar el DataFrame ajustando para visualización en Power BI
dataset_hashed.to_csv(
    'Datasets Procesados\\directorio_empresas_clusterizado_viz.csv', 
    sep=';',               # Separador de columnas
    decimal=',',           # Separador decimal
    index=False,           # Excluir índice
    encoding='utf-8-sig'   # Codificación compatible con Excel y Power BI
)


Datset para caracterización en Python

pd_merged.to_csv('Datasets Procesados\\directorio_empresas_final_clusterizado.csv', index=False)
